Decouple service implementations of Gradle modes
In Vintage, CC and IP, we need to use different implementations of some services, which are expected to be available in all modes.
