+++ platforms/core-configuration/configuration-cache/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.internal.cc.impl.ConfigurationCacheServices
+org.gradle.internal.cc.impl.BuildTreeControlServices
