+++ testing/architecture-test/src/changes/archunit-store/platform-coupling.txt
@@ -52,15 +52,17 @@ Class 'org.gradle.internal.buildconfiguration.tasks.DaemonJvmPropertiesAccessor'
 Class 'org.gradle.internal.buildconfiguration.tasks.DaemonJvmPropertiesAccessor' from platform 'core' imports class 'org.gradle.jvm.toolchain.internal.DefaultJavaLanguageVersion' from a forbidden platform 'jvm'
 Class 'org.gradle.internal.buildconfiguration.tasks.DaemonJvmPropertiesAccessor' from platform 'core' imports class 'org.gradle.jvm.toolchain.internal.DefaultJvmVendorSpec' from a forbidden platform 'jvm'
 Class 'org.gradle.internal.buildconfiguration.tasks.DaemonJvmPropertiesModifier' from platform 'core' imports class 'org.gradle.jvm.toolchain.JavaLanguageVersion' from a forbidden platform 'jvm'
-Class 'org.gradle.internal.cc.impl.BuildTreeModelControllerServices$VintageModelProvider' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.projectmodule.LocalComponentCache' from a forbidden platform 'software'
-Class 'org.gradle.internal.cc.impl.ConfigurationCacheServices$RemoteScriptUpToDateCheckerProvider$createRemoteScriptUpToDateChecker$1' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.modulecache.FileStoreAndIndexProvider' from a forbidden platform 'software'
-Class 'org.gradle.internal.cc.impl.ConfigurationCacheServices$RemoteScriptUpToDateCheckerProvider' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.ArtifactCachesProvider' from a forbidden platform 'software'
-Class 'org.gradle.internal.cc.impl.ConfigurationCacheServices$RemoteScriptUpToDateCheckerProvider' from platform 'core' imports class 'org.gradle.internal.resource.connector.ResourceConnectorFactory' from a forbidden platform 'software'
+Class 'org.gradle.internal.cc.impl.BaseGradleModeServices$BuildTree$createRemoteScriptUpToDateChecker$1' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.modulecache.FileStoreAndIndexProvider' from a forbidden platform 'software'
+Class 'org.gradle.internal.cc.impl.BaseGradleModeServices$BuildTree' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.ArtifactCachesProvider' from a forbidden platform 'software'
+Class 'org.gradle.internal.cc.impl.BaseGradleModeServices$BuildTree' from platform 'core' imports class 'org.gradle.internal.resource.connector.ResourceConnectorFactory' from a forbidden platform 'software'
+Class 'org.gradle.internal.cc.impl.ConfigurationCacheGradleModeServices$BuildTree' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.projectmodule.LocalComponentCache' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.ConfigurationCacheState' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ResolvableArtifact' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.ConfigurationCacheState' from platform 'core' imports class 'org.gradle.api.internal.artifacts.transform.TransformStepNode' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.ConfigurationCacheState' from platform 'core' imports class 'org.gradle.internal.resource.ResourceLocation' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.ConfigurationCacheState' from platform 'core' imports class 'org.gradle.internal.resource.TextResource' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.DefaultConfigurationCacheHost$DefaultConfigurationCacheBuild' from platform 'core' imports class 'org.gradle.internal.resource.TextFileResourceLoader' from a forbidden platform 'software'
+Class 'org.gradle.internal.cc.impl.IsolatedProjectsGradleModeServices$BuildTree' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.projectmodule.LocalComponentCache' from a forbidden platform 'software'
+Class 'org.gradle.internal.cc.impl.VintageGradleModeServices$BuildTree' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.projectmodule.LocalComponentCache' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.fingerprint.ConfigurationCacheFingerprintWriter' from platform 'core' imports class 'org.gradle.api.internal.artifacts.ivyservice.CacheExpirationControl$Expiry' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.fingerprint.ConfigurationCacheFingerprintWriter' from platform 'core' imports class 'org.gradle.internal.resource.TextResource' from a forbidden platform 'software'
 Class 'org.gradle.internal.cc.impl.initialization.AbstractInjectedClasspathInstrumentationStrategy' from platform 'core' imports class 'org.gradle.plugin.use.resolve.service.internal.InjectedClasspathInstrumentationStrategy$TransformMode' from a forbidden platform 'extensibility'
